breaking out proxy, environment objects; specs are still intermittently failing, but works in host app
