+++ spec/rspec-console/config_cache_spec.rb
@@ -21,24 +21,17 @@ describe RSpecConsole::ConfigCache do
 
   describe "#cache" do
     it "creates a proxy on first run" do
-      expect(Proxy).to receive(:new).and_call_original
+      expect(RSpecConsole::Proxy).to receive(:new).and_call_original
       config_cache.cache &config_block
     end
     it "delegates through proxy on repeated runs" do
       config_cache.cache &config_block
-      expect(Proxy).to_not receive(:new)
+      expect(RSpecConsole::Proxy).to_not receive(:new)
       config_cache.cache &config_block
     end
     it "defines method_missing method on RSpec.configuration's singleton class" do
       config_cache.cache &config_block
       expect(RSpec.configuration.respond_to?(:method_missing)).to eq(true)
-    end
-  end
-
-  describe "Proxy" do
-    it "records the original config" do
-      proxy = Proxy.new(::RSpec.configuration, ::RSpec.configuration)
-      expect(proxy.output).to eq(proxy.target)
     end
   end
 end
