+++ spec/rspec-console/environment_spec.rb
@@ -1,0 +1,9 @@
+require "spec_helper"
+
+describe RSpecConsole::Environment do
+  describe "#reset" do
+    it "caches rspec config while clearing env details" do
+
+    end
+  end
+end
