+++ spec/rspec-console/proxy_spec.rb
@@ -1,0 +1,17 @@
+require "spec_helper"
+
+describe RSpecConsole::Proxy do
+  let(:proxy) { RSpecConsole::Proxy.new(RSpec.configuration) }
+  it "records the original config" do
+    proxy.ordering_manager
+    expect(proxy.output).to eq(
+      [
+        {
+          method: :ordering_manager,
+          args: [],
+          block: nil
+        }
+      ]
+    )
+  end
+end
