+++ spec/spec_helper.rb
@@ -1,5 +1,6 @@
 $LOAD_PATH.unshift File.expand_path("../../", __FILE__)
 require "rspec-console"
+require "pry"
 
 RSpec.configure do |config|
   config.run_all_when_everything_filtered = true
